+++ packages/adapter-mongoose/src/utils.ts
@@ -1,42 +1,35 @@
-import type { RefreshTokenSchema, SessionSchema, UserSchema } from "lucia-sveltekit/types";
+import type { SessionSchema, UserSchema } from "lucia-sveltekit/adapter";
 
 export const convertUserDoc = (row: UserDoc): UserSchema => {
     const {
         _id: id,
         __v: _,
+        $__,
+        _doc,
         hashed_password,
         provider_id,
-        ...userData
+        ...attributes
     } = row;
     return {
         id,
         hashed_password,
         provider_id,
-        ...userData,
+        ...attributes,
     };
 };
 
 export const convertSessionDoc = (row: SessionDoc): SessionSchema => {
     const {
-        _id: _,
-        __v: _v,
-        access_token,
-        user_id,
+        _id: id,
+        __v: _,
+        user_id: userId,
         expires,
+        idle_expires: idleExpires
     } = row;
     return {
-        access_token,
-        user_id,
+        id,
+        user_id: userId,
         expires,
-    };
-};
-
-export const convertRefreshTokenDoc = (
-    row: RefreshTokenDoc
-): RefreshTokenSchema => {
-    const { _id, __v: _v, user_id, refresh_token } = row;
-    return {
-        user_id,
-        refresh_token,
+        idle_expires: idleExpires
     };
 };