+++ packages/lucia-sveltekit/src/adapter/index.ts
@@ -1,52 +1,65 @@
-import { SessionSchema, UserSchema } from "../types.js";
+export type UserSchema = {
+    id: string;
+    hashed_password: string | null;
+    provider_id: string;
+} & Lucia.UserAttributes;
+
+export type UserData = { id: string } & Required<Lucia.UserAttributes>;
+
+export type SessionSchema = {
+    id: string;
+    expires: number;
+    idle_expires: number;
+    user_id: string;
+};
 
 export interface Adapter {
-    getUserById: (userId: string) => Promise<UserSchema | null>;
+    getUser: (userId: string) => Promise<UserSchema | null>;
     getUserByProviderId: (providerId: string) => Promise<UserSchema | null>;
-    getUserByAccessToken: (accessToken: string) => Promise<UserSchema | null>;
+    getSessionAndUserBySessionId: (sessionId: string) => Promise<{
+        user: UserSchema;
+        session: SessionSchema;
+    } | null>;
     setUser: (
         userId: string | null,
         data: {
             providerId: string;
             hashedPassword: string | null;
-            userData: Record<string, any>;
+            attributes: Record<string, any>;
         }
-    ) => Promise<string>;
+    ) => Promise<UserSchema>;
     deleteUser: (userId: string) => Promise<void>;
     updateUser: (
         userId: string,
         data: {
             providerId?: string | null;
             hashedPassword?: string | null;
-            userData?: Record<string, any>;
+            attributes?: Record<string, any>;
         }
     ) => Promise<UserSchema>;
-    getSessionByAccessToken: (
-        accessToken: string
-    ) => Promise<SessionSchema | null>;
+    getSession: (sessionId: string) => Promise<SessionSchema | null>;
     getSessionsByUserId: (userId: string) => Promise<SessionSchema[]>;
     setSession: (
-        userId: string,
-        accessToken: string,
-        expires: number
+        sessionId: string,
+        data: {
+            userId: string;
+            expires: number;
+            idlePeriodExpires: number;
+        }
     ) => Promise<void>;
-    deleteSessionByAccessToken: (...accessToken: string[]) => Promise<void>;
+    deleteSession: (...sessionIds: string[]) => Promise<void>;
     deleteSessionsByUserId: (userId: string) => Promise<void>;
-    getUserIdByRefreshToken: (refreshToken: string) => Promise<string | null>;
-    setRefreshToken: (refreshToken: string, userId: string) => Promise<void>;
-    deleteRefreshToken: (...refreshToken: string[]) => Promise<void>;
-    deleteRefreshTokensByUserId: (userId: string) => Promise<void>;
 }
 
 export const getUpdateData = (data: {
     providerId?: string | null;
     hashedPassword?: string | null;
-    userData?: Record<string, any>;
-}) => {
+    attributes?: Record<string, any>;
+}): Partial<UserSchema> => {
     const rawData: Record<string, any> = {
         provider_id: data.providerId,
         hashed_password: data.hashedPassword,
-        ...(data.userData || {}),
+        ...(data.attributes || {}),
     };
     const result: Record<string, any> = {};
     for (const key in rawData) {