+++ packages/lucia-sveltekit/src/auth/request.ts
@@ -3,30 +3,22 @@ import type { Context } from "./index.js";
 import cookie from "cookie";
 import { Session } from "../types.js";
 
-type ParseRequest = (request: Request) => Promise<{
-    accessToken: string;
-    refreshToken: string;
-}>;
+type ParseRequest = (request: Request) => string;
 
 export const parseRequestFunction = (context: Context) => {
-    const parseRequest: ParseRequest = async (request) => {
-        const clonedReq = request.clone();
-        const cookies = cookie.parse(clonedReq.headers.get("cookie") || "");
-        const refreshToken = cookies.refresh_token || "";
-        const accessToken = cookies.access_token || "";
+    const parseRequest: ParseRequest = (request) => {
+        const cookies = cookie.parse(request.headers.get("cookie") || "");
+        const sessionId = cookies.auth_session || "";
         const checkForCsrf =
-            clonedReq.method !== "GET" && clonedReq.method !== "HEAD";
+            request.method !== "GET" && request.method !== "HEAD";
         if (checkForCsrf && context.csrfProtection) {
-            const origin = clonedReq.headers.get("Origin");
-            const url = new URL(clonedReq.url);
+            const origin = request.headers.get("Origin");
+            const url = new URL(request.url);
             if (!origin) throw new LuciaError("AUTH_INVALID_REQUEST");
             if (url.origin !== origin)
                 throw new LuciaError("AUTH_INVALID_REQUEST");
         }
-        return {
-            accessToken,
-            refreshToken,
-        };
+        return sessionId;
     };
     return parseRequest;
 };
@@ -35,9 +27,8 @@ type ValidateRequest = (request: Request) => Promise<Session>;
 
 export const validateRequestFunction = (context: Context) => {
     const validateRequest: ValidateRequest = async (request) => {
-        const { accessToken } = await context.auth.parseRequest(request);
-        if (!accessToken) throw new LuciaError("AUTH_INVALID_ACCESS_TOKEN");
-        const session = await context.auth.validateAccessToken(accessToken);
+        const sessionId = context.auth.parseRequest(request);
+        const session = await context.auth.validateSession(sessionId);
         return session;
     };
     return validateRequest;