+++ packages/lucia-sveltekit/src/client.ts
@@ -1,5 +1,6 @@
 import { getClientUser, getSSRUser } from "./user.js";
 import { LuciaError } from "./error.js";
+import { User } from "./types.js";
 
 export const signOut = async (redirect?: string): Promise<void> => {
     const user = getClientUser();
@@ -23,27 +24,7 @@ export const signOut = async (redirect?: string): Promise<void> => {
     if (result.message) throw new LuciaError(result.message);
 };
 
-export const refreshSession = async (): Promise<number> => {
-    const response = await fetch("/api/auth/refresh-session", {
-        method: "POST"
-    });
-    if (!response.ok) {
-        let result;
-        try {
-            result = await response.json();
-        } catch (e) {
-            console.error(e);
-            throw new LuciaError("UNKNOWN_ERROR");
-        }
-        throw new LuciaError(result.message);
-    }
-    const result = await response.json() as {
-        expires: number
-    }
-    return result.expires
-};
-
-export const getUser = () => {
+export const getUser = (): Required<User> | null => {
     if (typeof window === "undefined") {
         // server
         return getSSRUser();