+++ packages/lucia-sveltekit/src/types.ts
@@ -1,37 +1,17 @@
 import type { LuciaError } from "./error.js";
 
-export type User = Lucia.UserData & {
-    userId: string;
-    providerId: string;
-};
-
-export type UserSchema = {
-    id: string;
-    hashed_password: string | null;
-    provider_id: string;
-} & Lucia.UserData;
-
-export type SessionSchema = {
-    access_token: string;
-    expires: number;
-    user_id: string;
-};
-
-export type RefreshTokenSchema = {
-    refresh_token: string,
-    user_id: string
-}
-
-export type Tokens = {
-    accessToken: [string, string];
-    refreshToken: [string, string];
-    cookies: string[];
-};
+export type User = ReturnType<Lucia.Auth["context"]["transformUserData"]>;
 
 export type Session = {
+    sessionId: string;
     userId: string;
     expires: number;
-}
+    idlePeriodExpires: number,
+};
 
 export type Env = "DEV" | "PROD";
 export type Error = typeof LuciaError;
+
+export type GetSession = () => Session | null
+export type SetSession = (session: Session) => void
+export type ClearSession = () => void